moved the ModelFixer and Solution out to the top folder, as they can be used by other repair methods
