+++ src/hvl/projectparmorel/Solution.java
@@ -1,4 +1,4 @@
-package hvl.projectparmorel.general;
+package hvl.projectparmorel;
 
 import java.io.File;
 import hvl.projectparmorel.exceptions.DistanceUnavailableException;
